+++ typegen/evaluation/ipynb_evaluation_template.py
@@ -24,32 +24,13 @@ sys.path.append('./../../')
 
 from constants import Column, Schema
 from tracing import TraceDataCategory
+from typegen.evaluation.metric_data_calculator import get_total_completeness_and_correctness
 # -
 
 # Use the data file paths stored by the evaluate command.
 
 METRIC_DATA_FILE_PATH = "./../../evaluation_data/data.pytype"
 PERFORMANCE_DATA_FILE_PATH = "./../../evaluation_data/poetry.npy_pytype.npy"
-
-
-def get_total_completeness_and_correctness(metric_data: pd.DataFrame) -> tuple[float, float]:
-    """Gets the total completeness & correctness of a given metric data."""
-    completeness_column = metric_data[Column.COMPLETENESS]
-    correctness_column = metric_data[Column.CORRECTNESS]
-
-    total_completeness_count = completeness_column[~completeness_column.isna()].shape[0]
-    total_completeness_is_true_count = completeness_column[completeness_column].shape[0]
-    total_correctness_count = correctness_column[correctness_column].shape[0]
-
-    total_completeness = 0
-    if total_completeness_count > 0:
-        total_completeness = total_completeness_is_true_count / total_completeness_count
-    total_correctness = 0
-    if total_completeness_is_true_count > 0:
-        total_correctness = total_correctness_count / total_completeness_is_true_count
-
-    return total_completeness, total_correctness
-
 
 # +
 metric_data_template = pd.DataFrame(columns=Schema.Metrics.keys())
